requirements.txt
streamlit
google-generativeai
scikit-learn
python-dotenv